Fix compilation errors in LibraryScreen and MainActivity

+++ app/src/main/java/com/example/musicdownloader/ui/PlaylistScreen.kt
@@ -134,7 +134,7 @@ fun PlaylistRow(playlist: Playlist, onClick: () -> Unit) {
 @Composable
 fun PlaylistDetailScreen(
     viewModel: MusicViewModel,
-    playlistId: Long,
+    playlistId: Int,
     playlistName: String,
     onBack: () -> Unit,
     onSongClick: (String) -> Unit
